+++ sayaya-ui/src/net/sayaya/ui/regacy/shape/impl/Line.java
@@ -6,18 +6,18 @@ import net.sayaya.ui.regacy.data.Point;
 import net.sayaya.ui.regacy.shape.HasStroke;
 
 public class Line extends ShapeInstance<Line> implements HasStroke {
-	private final Integer[] start;
-	private final Integer[] end;
+	private final Double[] start;
+	private final Double[] end;
 	private String color;
 	private double borderWidth;
-	public Line(Integer[] start, Integer[] end) {
+	public Line(Double[] start, Double[] end) {
 		this.start = start;
 		this.end = end;
 	}
 	
 	public Line(Point<? extends Number, ? extends Number> start, Point<? extends Number, ? extends Number> end) {
-		this.start = new Integer[] {start.getX().intValue(), start.getY().intValue()};
-		this.end  = new Integer[] {end.getX().intValue(), end.getY().intValue()};
+		this.start = new Double[] {start.getX().doubleValue(), start.getY().doubleValue()};
+		this.end  = new Double[] {end.getX().doubleValue(), end.getY().doubleValue()};
 	}
 	@Override
 	public boolean checkIn(double x, double y) {
